api/requirements.txt
azure-functions
pydantic
azure-cosmos
azure-storage-blob
